# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IT Asset Inventory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Asset {
  id: number;
  assetNumber: string;
  name: string;
  category: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  status: string;
  /** @nullable */
  currentUser?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt?: string | null;
}

export interface AssetInput {
  /** @minLength 1 */
  assetNumber: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  brand?: string;
  model?: string;
  serialNumber?: string;
  /** @minLength 1 */
  status: string;
  currentUser?: string;
  location?: string;
  notes?: string;
}

export interface AssetUpdate {
  /** @minLength 1 */
  assetNumber?: string;
  /** @minLength 1 */
  name?: string;
  category?: string;
  brand?: string;
  model?: string;
  serialNumber?: string;
  status?: string;
  currentUser?: string;
  location?: string;
  notes?: string;
}

export interface HistoryEntry {
  id: number;
  assetId: number;
  /** @nullable */
  assetName?: string | null;
  /** @nullable */
  assetNumber?: string | null;
  changedAt: string;
  changedBy: string;
  changeType: string;
  /** @nullable */
  fieldChanged?: string | null;
  /** @nullable */
  oldValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Note {
  id: number;
  assetId: number;
  content: string;
  createdAt: string;
  createdBy: string;
}

export interface NoteInput {
  /** @minLength 1 */
  content: string;
  /** @minLength 1 */
  createdBy: string;
}

export type DashboardSummaryByStatusItem = {
  status: string;
  count: number;
};

export type DashboardSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface DashboardSummary {
  totalAssets: number;
  disabledCount: number;
  trashedCount: number;
  byStatus: DashboardSummaryByStatusItem[];
  byCategory: DashboardSummaryByCategoryItem[];
  recentlyAdded: number;
}

export type ListAssetsParams = {
search?: string;
category?: string;
status?: string;
};