# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IT Asset Inventory API
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  assetNumber: string;
  name: string;
  category: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  status: string;
  /** @nullable */
  currentUser?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  deletedAt?: string | null;
}