# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IT Asset Inventory API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryByCategoryItem } from './dashboardSummaryByCategoryItem';
import type { DashboardSummaryByStatusItem } from './dashboardSummaryByStatusItem';

export interface DashboardSummary {
  totalAssets: number;
  disabledCount: number;
  trashedCount: number;
  byStatus: DashboardSummaryByStatusItem[];
  byCategory: DashboardSummaryByCategoryItem[];
  recentlyAdded: number;
}